requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.5
face_recognition==1.3.0
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
opencv-python==4.8.1.78
